Fix RGB buffer size (from 11 bytes to 12)
The make_rgb() function generated a string in the "r,g,b" format.
Since the individual bytes are 8-bit, the longest possible generated
string is "XXX,YYY,ZZZ" - which is 11 bytes _without_ the NUL-terminator.
To store a proper, NUL-terminated string, we need 12 bytes.

The RGB buffer size was set to 11, which meant that in the longest-case,
the code wrote one byte past the buffer. This could lead to a segfault.