Adding brief details on 2 other ways to install
This package is exactly what I was looking for.  Thanks for creating it.  I found it via conda, so was suprised to see it was not mentioned as an install option (nor homebrew- i tested brew aha on Catalina, it worked, and I tested conda aha on Ubuntu 18 and 20 and it worked).  These routes will certainly make the tool more approachable.  Thanks again--